Text.ParserCombinators.Parsec

Description

Parsec compatibility module

Documentation

module Text.ParserCombinators.Parsec.Prim

module Text.ParserCombinators.Parsec.Combinator

module Text.ParserCombinators.Parsec.Char

data ParseError

The abstract data type ParseError represents parse errors. It provides the source position (SourcePos) of the error and a list of error messages (Message). A ParseError can be returned by the function parse. ParseError is an instance of the Show and Eq classes.

errorPos :: ParseError -> SourcePos

Extracts the source position from the parse error

data SourcePos

The abstract data type SourcePos represents source positions. It contains the name of the source (i.e. file name), a line number and a column number. SourcePos is an instance of the Show, Eq and Ord class.

type SourceName = String

type Line = Int

type Column = Int

sourceName :: SourcePos -> SourceName

Extracts the name of the source from a source position.

sourceLine :: SourcePos -> Line

Extracts the line number from a source position.

sourceColumn :: SourcePos -> Column

Extracts the column number from a source position.

incSourceLine :: SourcePos -> Line -> SourcePos

Increments the line number of a source position.

incSourceColumn :: SourcePos -> Column -> SourcePos

Increments the column number of a source position.

setSourceLine :: SourcePos -> Line -> SourcePos

Set the line number of a source position.

setSourceColumn :: SourcePos -> Column -> SourcePos

Set the column number of a source position.

setSourceName :: SourcePos -> SourceName -> SourcePos

Set the name of the source.