Text.Parsec.Language

Description

A helper module that defines some language definitions that can be used to instantiate a token parser (see Text.Parsec.Token).

Documentation

haskellDef :: LanguageDef st

The language definition for the Haskell language.

haskell :: TokenParser st

A lexer for the haskell language.

mondrianDef :: LanguageDef st

The language definition for the language Mondrian.

mondrian :: TokenParser st

A lexer for the mondrian language.

emptyDef :: LanguageDef st

haskellStyle :: LanguageDef st

This is a minimal token definition for Haskell style languages. It defines the style of comments, valid identifiers and case sensitivity. It does not define any reserved words or operators.

javaStyle :: LanguageDef st

This is a minimal token definition for Java style languages. It defines the style of comments, valid identifiers and case sensitivity. It does not define any reserved words or operators.

type LanguageDef st = GenLanguageDef String st Identity

data GenLanguageDef s u m

The GenLanguageDef type is a record that contains all parameterizable features of the Token module. The module Language contains some default definitions.